draft(animations): draft commit ive gave up this rebuilding... too hard!
