+++ PCL.Neo/Helpers/Animation/AnimationChain.cs
@@ -6,13 +6,12 @@ using IAnimation = PCL.Neo.Animations.IAnimation;
 
 namespace PCL.Neo.Helpers.Animation
 {
-    public class AnimationChain(Animatable control) : IAnimationChain, IDisposable
+    public class AnimationChain(Animatable control) : IDisposable
     {
-        public Animatable Control { get; init; } = control;
-        public List<IAnimation> Animations { get; } = [];
-        public bool IsLoop { get; init; }
-        public bool IsComplete { get; set; }
-        public CancellationTokenSource CancellationToken { get; } = new();
+        internal Animatable Control { get; } = control;
+        internal List<IAnimation> Animations { get; } = [];
+        public bool IsComplete { get; internal set; }
+        internal CancellationTokenSource CancellationToken { get; } = new();
 
         public void Cancel()
         {
