+++ PCL.Neo/ViewModels/HomeViewModel.cs
@@ -1,14 +1,14 @@
+using Avalonia.Media.Imaging;
+using Avalonia.Platform;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using PCL.Neo.Models.User;
 using PCL.Neo.Services;
 using PCL.Neo.ViewModels.Home;
-using System;
-using System.Threading.Tasks;
-using Avalonia.Media.Imaging;
-using Avalonia.Platform;
 using SkiaSharp;
+using System;
 using System.IO;
+using System.Threading.Tasks;
 
 namespace PCL.Neo.ViewModels;
